feat(repo-explorer): extracted GitCommit type to models.ts
