+++ src/models.ts
@@ -39,3 +39,14 @@ export type BranchIdentity = {
  * Объект, описывающий файл или директорию
  */
 export type GitFile = { type: string; name: string };
+
+/**
+ * Основная информация о коммите
+ */
+export type GitCommit = {
+    message: string;
+    login?: string;
+    avatarUrl?: string;
+    name?: string | null;
+    date: string;
+};
